Update logic to display the TagConfigRows correctly.
